GEODE-7476: Fix Windows break *again* - Linux fix broke Windows, after previous Windows fix broke Linux

+++ cppcache/src/statistics/StatArchiveWriter.cpp
@@ -222,7 +222,7 @@ void ResourceInst::writeSample() {
         wroteInstId = true;
         writeResourceInst(dataOut, id);
       }
-      dataOut->writeByte(i);
+      dataOut->writeByte(static_cast<int8_t>(i));
       writeStatValue(stats[i], delta);
     }
   }
